moving losses to main package
