+++ scripts/machine_translation/train_transformer.py
@@ -45,11 +45,12 @@ import mxnet as mx
 from mxnet import gluon
 import gluonnlp as nlp
 
+from gluonnlp.loss import SoftmaxCEMaskedLoss, LabelSmoothing
 from gluonnlp.model.translation import NMTModel
 from gluonnlp.model.transformer import get_transformer_encoder_decoder, ParallelTransformer
 from gluonnlp.utils.parallel import Parallel
 from translation import BeamSearchTranslator
-from loss import SoftmaxCEMaskedLoss, LabelSmoothing
+
 from utils import logging_config
 from bleu import _bpe_to_words, compute_bleu
 import dataprocessor
