+++ src/gluonnlp/loss/label_smoothing.py
@@ -16,46 +16,15 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Loss functions."""
+
+"""Label Smoothing"""
+
+__all__ = ['LabelSmoothing']
 
 import numpy as np
 import mxnet as mx
 from mxnet.gluon import HybridBlock
-from mxnet.gluon.loss import SoftmaxCELoss
 
-__all__ = ['SoftmaxCEMaskedLoss', 'LabelSmoothing']
-
-
-class SoftmaxCEMaskedLoss(SoftmaxCELoss):
-    """Wrapper of the SoftmaxCELoss that supports valid_length as the input
-
-    """
-    def hybrid_forward(self, F, pred, label, valid_length): # pylint: disable=arguments-differ
-        """
-
-        Parameters
-        ----------
-        F
-        pred : Symbol or NDArray
-            Shape (batch_size, length, V)
-        label : Symbol or NDArray
-            Shape (batch_size, length)
-        valid_length : Symbol or NDArray
-            Shape (batch_size, )
-        Returns
-        -------
-        loss : Symbol or NDArray
-            Shape (batch_size,)
-        """
-        if self._sparse_label:
-            sample_weight = F.cast(F.expand_dims(F.ones_like(label), axis=-1), dtype=np.float32)
-        else:
-            sample_weight = F.ones_like(label)
-        sample_weight = F.SequenceMask(sample_weight,
-                                       sequence_length=valid_length,
-                                       use_sequence_length=True,
-                                       axis=1)
-        return super(SoftmaxCEMaskedLoss, self).hybrid_forward(F, pred, label, sample_weight)
 
 # pylint: disable=unused-argument
 class _SmoothingWithDim(mx.operator.CustomOp):
